+++ src/test/java/umc/cockple/demo/domain/chat/service/query/DirectChatRoomQueryServiceTest.java
@@ -63,7 +63,7 @@ class DirectChatRoomQueryServiceTest {
     @BeforeEach
     void setUp() {
         chatConverter = new ChatConverter();
-        chatUnreadQueryService = new ChatUnreadQueryService(messageReadStatusRepository);
+        chatUnreadQueryService = new ChatUnreadQueryService(messageReadStatusRepository, chatConverter);
         imageUrlResolver = new ImageUrlResolver(fileService);
         directChatRoomQueryService = new DirectChatRoomQueryService(
                 chatRoomRepository,
